feat: add onResourcesCount to API
