+++ packages/channels/src/model/index.ts
@@ -26,7 +26,6 @@ export * from './endpoint';
 export * from './endpoints-info';
 export * from './endpoints-options';
 export * from './kubernetes-providers-info';
-export * from './kubernetes-resource-count';
 export * from './kubernetes-troubleshooting';
 export * from './openshift-types';
 export * from './pod-logs-chunk';
@@ -37,7 +36,6 @@ export * from './resource-details-info';
 export * from './resource-details-options';
 export * from './resource-events-info';
 export * from './resource-events-options';
-export * from './resources-count-info';
 export * from './target-ref';
 export * from './update-resource-info';
 export * from './update-resource-options';
